spt: Use 'long' for return value from block/net read/write
Consistent with e6f0684b1d2fc954f02011027ac23c10ac311d7a.

+++ bindings/spt/block.c
@@ -58,7 +58,7 @@ solo5_result_t solo5_block_read(solo5_off_t offset, uint8_t *buf, size_t size)
     if(offset > (block_capacity - block_size))
         return SOLO5_R_EINVAL;
 
-    int nbytes = sys_pread64(block_fd, (char *)buf, size, offset);
+    long nbytes = sys_pread64(block_fd, (char *)buf, size, offset);
 
     return (nbytes == (int)size) ? SOLO5_R_OK : SOLO5_R_EUNSPEC;
 }
@@ -79,7 +79,7 @@ solo5_result_t solo5_block_write(solo5_off_t offset, const uint8_t *buf,
     if(offset > (block_capacity - block_size))
         return SOLO5_R_EINVAL;
    
-    int nbytes = sys_pwrite64(block_fd, (const char *)buf, size, offset);
+    long nbytes = sys_pwrite64(block_fd, (const char *)buf, size, offset);
 
     return (nbytes == (int)size) ? SOLO5_R_OK : SOLO5_R_EUNSPEC;
 }
